chore(wc): move v2 code into root saga
